requirements.txt
Django==1.9.7
django-nose==1.3
django-tastypie==0.12.1
python-dateutil==2.4.2
python-mimeparse==0.1.4
six==1.9.0
django-coverage==1.2.4
coverage==3.7.1
python-coveralls==2.5.0
coveralls==0.5
